Have display page UC use PageRetriever and PageContentModifier
PageRetriever and implementations copied from old repo and somewhat
modified. Currently unclear to me when which implementation should
be used, and which ones we actually need. Did not move over the
ApiBasedPageRetriever yet, as it has additional depenendecies.

Introduced PageContentModifier class, which holds modified
code from the old WikiRip::getProcessedContent.

For both of these I dropped the action parameter, as it appears
to always be "render" for this UC. We can re-add this later if
needed, probably in a different manner.