Refactor RequestValidator
RequestValidator demonstrates how to set up a class that implements the new InstanceValidator interface. The Validate method is so generic it could be used in a base class or a trait.
Added RequiredFieldValidator for checking if a field is empty.
Added ValidationConstraint to add field and class information to ConstraintViolation instances.