Add demo of paged form
Add a browser-testable version of the paged form.

Add npm compilation with browserify to package.js.
Ignore compiled JavaScript with .gitignore
Properly import jQuery with browserify-shim to use in a browser context
where jQuery is also used for other stuff. Import it as a node module
when running unit tests.