Throw AuthorizationUpdateException in DoctrineAuthorizationUpdater
Test not implemented yet, as I do not know to do this in a nice
way. The read of the Donation should still work, while the write
should throw an exception. We could create a fake that mimicks
the behaviour and implements the whole of EntityManagerInterface,
though that is really not nice.