Add initial state to form
It's now possible to initialize the form values with initial values
(which will be generated by the backend in a future commit).

Improved validation code to for easier handling of initial state.
Validation needs to know about initial state to allow for invalid values
as initial state (e.g. empty fields).

This is one step for task https://phabricator.wikimedia.org/T129227