Split initialization and event handling code
Created the `store_update_handling` module contains functions to
dispatch the update events from the store to various receivers. This
makes the connection of components, view handlers and update events
testable. It also avoids repeated builerplate code between the
DonationForm template and the HOWTO and makes the HOWTO more clear.