The AddDonation UC now writes its token to a cookie
Added a test verifying this from an as high level as possible,
to not bind to implementation details and make sure the combination
of the routes works correctly. I'm not sure we're already
providing the interface we want to the clients with regards to
the update token, so that might still need further work.