Add element slide view handler
Add view handler for showing and hiding elements when a value in the
state matches a regular expression.

This is for replacing older functionality and a preparation for the bank
data fields on the 2nd page of the form.

It is now possible for view handlers to access single properties nested
in the state object.