Added initial version of DoctrineMembershipApplicationRepository
The data field still needs to be set on write, and the read method
has not been implemented at all yet.

Though by all means merge away. The code here is complete in the
sense of that it does not fail, is tested, and is clean.

The tests here are an iteration on the ones for the Donation repository,
and I'm liking these a lot more now. They are nicer and more complete.

Part of https://phabricator.wikimedia.org/T123117
and https://phabricator.wikimedia.org/T130199