Translate frontend and validation field names
Translate frontend and validation field names from German to English.
The field names in the data field must remain in German for backwards
compatibility with the backend and export scripts.

In the future, we could just duplicate the information in the data
object with english field names to gradually move to all-english names.