Improve donation failure template
Changed factory code to use the DonationForm template as a page instead
of displaying it directly (which would result in missing surrounding
HTML).
Added comment to make the form name configurable for A/B testing.

Added debugging code to have a peek at the validation errors coming from
the server.
Added TODOs to the DonationForm template.