Make comment private when changing status from moderation or cancelled
Finishes https://phabricator.wikimedia.org/T137328

Please verify the logic is correct. What I did is check if the status is either
moderation or cancelled, as opposed to checking if the status is not external
incomplete as described in the ticket. I suspect these are equivalent, though
am not sure as the later is not very explicit.