Added profiling decorators
Only for CommentFinder, PageRetriever and Guzzle Client so far.
The later two are presumably the most important ones to profile.

Note that the generic decorator building approach relies on PHPUnits
mocking API, which internally builds a string and evals it. So this
is presumably something we do not want to use in the production code
when it is not initialized by the dev entry point. (This PR makes
sure that does not happen.)

Follow up things to do:

* Decorate ALL the interesting things with profilers
* Do something about the global access in the factory
* Move the decorator building code elsewhere

Can already be merged as far as I'm concerned though.