Moved logger construction out of factory
Only behaviour change for now is that the tests will no longer use a logger.
We can set one up in TestEnvironment if we want to though.

Like this we can set up different logging for production and for the dev
entry point. Presumably we want those to write to different loggs. Also,
right now the non-dev entry point is logging every request, which we presumably
do not want to do.