Add Validation states
Apart from the `OK` and `ERR` states we had `INCOMPLETE`. `Now they are
all bundled up in a value object.

For the bank data validation the state `NOT_APPLICABLE` was added to
avoid sending data to the server and changing the validation state when
a different payment method thank direct debit is selected.