Improve resilience for URLs in comments
Address the concerns in
https://github.com/wmde/FundraisingFrontend/pull/621#issue-169557470

While looking at the code, it was not quite clear if certain code paths
could be tested in unit test cases. Further research hinted at the fact
that the code paths might not be testable:

https://www.reddit.com/r/lolphp/comments/4sgcmj/on_seriously_malformed_urls_parse_url_may_return/
http://stackoverflow.com/questions/38956582/can-parse-url-ever-detect-a-malformed-url-when-combined-with-a-regular-expressio

The point of the TextPolicyValidator is to detect URLs but be lenient
for URL-lookalikes and non-existing URLs. So instead of `parse_url`,
which required 3 different code paths that could not be tested properly,
`filter_var` is now used for URL detection.