Set different logger names for entry points
The logger "names" (channels) now designate what entry point was used.
Until we have different channels for different parts of the application
(db_layer, use_case, service_name, etc) the entry point adds useful
information.

Using underscore instead of dots for the file name avoids confusion when
looking for the source of an error and when grepping through the files.