Handle incomplete data in donations
Created test data in IncompleteDoctrineDonation that simulates purged
payment data and missing tracking information. Test if the data can be
loaded successfully none the less.

The script load_all_donations is was used to check if all the donations
in the live database can be loaded without fatal errors.

There were ~70 donations where the loading failed due to donation amount
being NULL or an empty string. Those entries will be updated in the
database and not handled by code.