Add explicit 404 presenter
The error message was not localizable and contained the HTTP request
method and the word "route" which can be considered jargon.

Also, every 404 error was logged with a stack trace.

This patch adds a localizable general message and does not log 404
errors to the app error log (the errors will still show up in the web
server logs).