Server-side tracking for mobile paypal donations
Mobile paypal donations are anonymous and go straight to donation/add
route which redirects to paypal. That means we don't have Piwik tracking
data which is needed by the backend. That means we have to simulate the
piwik tracking on the server.

ServerSideTracker, its implementations and PageViewTracker were taken
from the old fundraising app.

The different tracking behavior, depending on the payment type would
lead to an increasing number of tests, this should be refactored some
time in the future.