Allow setting donor name type for new donations
addressType (designating both the name type of the donor and if there
should be a donor at all) can now be passed in the request from the
banner and will be selected on the 2nd form page. When the value is
"anonym", no address fields will be presented on the 2nd page.