Reorganize email configuration
Membership-related emails (or rather, emails pertaining to the main
organization (WMDE) can now have a different sender/receiver than emails
that pertain to the sub-organization (Wikimedia Fördergesellschaft).

In the future, the `contact_info` configuration could be used for
different use cases (contact, subscription, etc), but at the moment
that's not needed and  would be too much duplication in the
configuration.