Don't change value of elements in focus
If a store value is set in a component while the user  edits it, the user will be confused. To avoid that, we only change values in components that aren't in focus. The "new" value from the store will be list anyway because the blur event will trigger an "onChange" action that sets the new value in the store.