Added transfer code generator with checksum
First step towards https://phabricator.wikimedia.org/T170266

Has checkstum and 8 characeters instead of 7 to add some extra entropy.
Like this we have 150 billion combinations, which should give us enough
room for adding a bad word filter.

Added a dash in the center of the thing for readability.

Did not include the prefix in this class as it is not clear to me if
it should go in there or not. If it depends on something else, such
as the type of donation, then probably not.