Mark technical debt of cat17 in JavaScript
The developers of the cat17 skin did not respect the architecture
of the 10h16 skin, but copy-pasted and edited some of its code.

Violations of the 10h16 architecture are marked with "fixme".
Possible improvements are marked with "TODO".

As long as the functionality works, the code does not need to be
changed, but for a model how data flow and event handling *should* work,
please refer to 10h16 and **do not** add more jQuery code. Instead,
follow the pattern of:

* state is separated from the DOM via redux
* state changes are triggered by actions, which can be called from DOM
  events or AJAX callbacks.
* DOM changes are done with view handers that react to state changes,
  **not** to DOM events.

skins/cat17/src/scripts/main.js could/should be completely refactored.