Use new donor validation code from the Donation BC
https://phabricator.wikimedia.org/T180936

Tests are failing since some or the error names and field names changed.
I'm not finding where these are consumed, if they are at all, so am not
sure if this needs to be updated or not. At least one change, field name
from "company" to "companyName" seems to make things better, as here the
name of the get parameter is "companyName".