Hide detail elements properly
Use `display: none` instead of `opacity` and `height` for hiding detail
elements. The latter were vestiges of trying to animate showing and
hiding the elements but since the animation was commented out anyways by
the 3rd party, we can use display and avoid invisible overlapping
elements that obstruct the submit button.