Set header element class during scrolling
The header status bar becomes only visible (with the CSS class `active`)
after a certain scroll position which messes with the calculation of the
final scroll offset.

This change introduces the `scrolling` CSS class that is only active on
an element when scrolling iwth the `scrollTo` method. By giving it the
same CSS properties as `active`, the element has the proper height
before the offset is calculcated.