Expose database container on separate network
Add a new network name `fundraising_network` to the docker-compose
configuration. This network name can be used to access the database from
other Docker containers, e.g. the MySQL client or docker-compose.yml
file of the Fundraising Operation Center.

Add documentation on how to access the database. The example with the
docker-compose.db.yml override file is unrelated to the network change
and just an example how to expose the database port.