Enable parallelized run of CI tasks
Add unique names to each container that is run with docker-compose.

Document how to run CI tasks in parallel.

We don't use parallel execution for Travis, because the `make` command
on Travis does not support group the process output by target
(--output-sync), which would make the log file harder to read.