Unify placeholder color across browsers in 10h16
The placeholder color of text fields was slightly ligther than the one
of dropdown fields on Firefox. This is because Firefox applies a [0.54
opacity](https://developer.mozilla.org/de/docs/Web/CSS/::-moz-placeholder)
to the placeholder element by default.

The differnet colors were very visible on the new contact form, if you
looked very closely the colors were also visible for the "Title" field
in the donation and membership forms.