Move some Silex routes to controller classes
This example shows how the anonymous functions from the routes.php files
can be moved into controller methods. This is a preparation to move to
Symfony, where all routing actions must be placed in controller methods.

This example also shows, how a RouteHandler class can be converted into
a controller class.

The new FundraisingFactoryServiceProvider enables the Silex Application
container to automatically inject the FunFunFactory into all controller
classes.

In the future, the FundraisingFactoryServiceProvider might also provide
the use case classes to the controllers, to make it more clear what the
controller is actually about and to further reduce the access surface.