Add minimal Vue.js setup for update form
Add a proof of concept setup that replaces the form with its own Vue-js
rendered element (not the whole form yet, just a custom component).

Add a basic webpack dev configuration, for production it needs to be
integrated with the gulp file and needs to be split into prod and dev
parts.