Add StreamOpener for bucket log
The Stream opener can recursively create file paths before opening
the file.

The ErrorLoggingStreamOpener guarantees a stream resource to be
returned. If the bucket log resource can't be opened, it logs an error
and opens a "dummy" (in memory) resource that will be lost after the
script ends.
See https://stackoverflow.com/q/52116166/130121 for a discussion about
why I chose `php://memory` as a fallback resource.