Use separate entry point for Vue
After many failed attempts to export a Vue component inside
src/app/main.js and use it with skins/cat17/src/scripts/donationForm.js,
we decided to create a separate entry point and output file.

This is a first working demo of a Vue component that's integrated into
the Redux store. It's unfortunate that it has to use a gloabl variable
for the store, but we could not make any other solution work.