Fix icons part of link menu
The arrow icon next to each link in a menu broke some time ago and it went unnoticed. A new set of icons
was generated in #1338 to add a new envelope icon. That unintentionally fixed the broken arrow icon.
The arrow icon seemed out of place in one 'back link', it was too far on the right compared to the text,
and didn't look too good in the rest of the links on desktop, therefor we removed it in #1353.
That introduced a bug, which broke the facebook and twitter icons on a menu which used to have the arrow icon.
This PR reverts the removal of that css and keeps the arrow icon only in mobile view, to indicate the menu is clickable.