Testing validity indication classes
The "valid" class shows if a bank data field was filled. In the future
we might rename it to "filled".

The "invalid" class now reflects the `isValid` property, which can have
three values: true, false and null. The latter means "not enough data to
determine validity".

Use ValidationState constants in validity reducer and change how
validity value is determined: State `INCOMPLETE` leads to `null` value.