Ignore bank data validity for non-direct debit
As a consequence from the removal of NOT_APPLICABLE, there was a bug
that prevented form submission if the bank data validation had an error
and the user switched to a different payment type.

This change acknowledges the fact that we can ignore the bank data
validity value in the section validity aggregate if the payment type is
not direct debit.