Move height adjustment to view handler
Context
The form fields are absolutely positioned in big viewports. Therefore,
their min-height needs to be adjusted so their surrounding container
flows around them.

This patch moves the adjustment from DOM change handlers on input fields
to a view handler that reacts to store events. This fixes errors that
pop up when the asynchronous validation is not fast enough or when the
store values change without input (i.e. pressing the submit button).