Introduce domain-specific date class
Campaign dates built by the CampaignBuilder were always in the UTC time
zone (because of the time zone conversion between the time zone
form the configuration file). On servers where the default time zone was
not UTC, this lead to errors in Campaign::isExpired.

The solution to this problem is introducing the CampaignDate class that
enforces a UTC time zone and does time zone conversion when given a
string and time zone.