Simplify store data model
Drop dataEntered value. In the old frontend code, where the values were
also put in the store and values could also come from the server, it was
necessary to track the user interaction. With Vue, we donÄt teack that
any more and don't need dataEntered.
Use Enum for Validity.

This replaces #1437