fix division by zero in UI validation
The paymenIntervalInMonths value being 0 is currently not reproducable
as its initial value is -1 and the UI validation prevents it from entering
a state where it could become 0.

This additional check will prevent the interval field from being validated
before it was selected because there is a check now for its correct
default value (-1).

https://phabricator.wikimedia.org/T215163