Add entry point for error messages
As a first proof of concept, add an entry point for displaying error
messages, showing the data exchange between the Twig-generated HTML
markup and the vue entry point.

This Proof-of-Concept shows the following areas where further refinement
is needed:
* A better way to pass the translation map directly to Vue and use a
function in templates, instead of adding a `messages` property to each
component.
* A generalized way to pass data to the vue instance (avoid the
boilerplate of checks, plus add types (interfaces) to the data model to
be more explicit about what data is required (and optional).
* A better way to build layouted apps. Currently, the `Error` page
component is standalone. A better way would be to have a root `App` with
a slot (its template consisting of layout components) and passing it the
current "Page Component" in the render function of the entry point.
* A convention to organize entry points, their "Page component" and
their data model definition.