Move validation route code to classes
In preparation for moving to Symfony, move validation route code to
Controller classes. Use JsonResponse instead of $app->json.
See https://phabricator.wikimedia.org/T198484

Also introduce route names, to be able to reference the validation
routes in the rest of the application code without binding to concrete
URLs and without using strings that might get out of sync.
This commit does **not** replace the references in the code with the
constants.