Move URL endpoints to template context
Change all templates where hardcoded URLs are in data attributes to URLs
from the template context. This is needed in the laika skin so we can
use the `_context` variable to pass data from PHP to JavaScript.

Add the URLs in all presenters/template contexts that need/use them.

This commit does *not* move all urls to the template context, there are
still some URLs of routes in the template. Future commits can move them
to Routes::getNamedRouteUrls

Move membership application form route to controller, because this
change is touching the code.