Improve localization string structure
The Laika skin will now have its own messages file in order to ensure
that only strings which are actually used by the skin are loaded by the
client. Previously, old and unused strings from the previous skins was
also loaded alongside new strings which was somewhat messy. This
approach also means we will not have to do a clean-up of the messages
files once the old skins will be removed, as the new laika_messages.json
will only contain strings which we actually need.