Make empty JCF dropdowns compatible with Vue
Our validation was triggered by blur events, which were triggered
prematurely by JCF, leading to erronous validation when the user
selected the dropdown with an empty default value. Now the dropdown
fields sned the value only when it's either not empty or there was a
previous value.

This fixes https://phabricator.wikimedia.org/T225481