Initialize membership address store
When the server provides address data from a previous donation, write it
to the store first and then initialize the Address component data with
it, instead of passing it down with props.

This makes the component code more clean and allows an instant "submit"
of the address page instead of having to go into each field to trigger
its validation.