Protect payment page against race conditions
Don't submit the page until the store is not validating any more.

The logic of "check if validation is going on, if it is, trigger callback
when validation is finished" is the same on all donation and membership
pages and should be moved to its own class in subseqent commits.