Separate static page data and privacy policy opt out
Both base.html and privacy_policy.html are the same but are needed because the route can load both of them.
There is one Vue component for all static pages that displays the content.
For Privacy Policy opt out there is an additional Vue component which is loaded conditionally.
That component loads the tempalte that was replaced in the cat17 solution.