Add payment summary to address page
Add the PaymentSummary component and display it, depending if the user
came from a banner or not.

Introduce the has-outside-border class that aligns its inner text with
the rest of the text on the page and puts the border outside of the
flow. This will need to be adjuested/tested for mobile (See #1625).