Log request parameters when CC notification fails
We're passing the incoming parameters of the credit card notification
unvalidated to the credit card request library (micropayment). It
validates the parameters and will throw an exception if important
parameters are missing. To avoid losing the context (values in other,
valid parameters), we log the query string when an error occurs.