Refactor Legacy Parameter Reading
Consolidate reading deprecated params and logging if they are found.
This will make it easier in the future to remove the deprecated
parameters and replace the FallbackRequestValueReader with a scalar
default value.

Change how the old parameters are converted to be more
backwards-compatible: Add support for float values and German-formatted
decimal numbers.

Add "Sunset Tests" to force us to remove the feature in the future.