Temporarily use composer and PHP of Travis
When using our own container for running "composer install", it's
missing a GitHub access token and fails. Trying to get around that by
using Travis' own composer which should have a token. This reintroduces
the problem of PHP versions getting out of sync between Travis and our
containers. In the long run, our container should have its own GitHub
access token, stored as a Travis secret. When that is done, revert this
commit.