Remove FundraisingStore
Add `DoctrineFactory` to create Doctrine instance with mappings and event
subscribers from all contexts.

Split creation of EntityManager in two phases to be able to use it to
recreate the database schema with the new `SchemaCreator` class.

Adapt AddressChange code to changes in AddressChange

Change Doctrine setup in `cli-config.php` for using the doctrine CLI.