Implement simple translator interface
Since we don't need the full functionality of Symfony translator, create
a simple interface for translation. Replace most occurrences of the
Symfony translation instance with the new implementation.

Switch test case implementations of
DonationConfirmationMailSubjectRendererTest and
MembershipConfirmationMailSubjectRendererTest, their methods were in the
wrong place.

Deprecate JsonStringReader, its functionality should be in the content
CI.