Move tracking storage to event handler
Created `RegisterTrackingData` event handler with code from
`Bootstrap.php`.

Removed the `tracking` POST parameter because that was coming from a
version of the app where the tracking parameters were preserved between
multiple form pages.

Fixed "find cookie" function calls in CampaignCookieTest - that was accidentally working, but now the order of cookies in the array has changed.