Replace deprecated Symfony\Component\HttpKernel\Client
Version 3 of the HttpKernel package has deprecated the `Client` class,
but does not have a ready-to-use replacement. Only HttpKernel in version
4.4 implements the HttpKernelBrowser class that is a replacement. Since
we can't use Version 4 with Silex, we use a self-implemented copy.

After the Symfony migration we should use the "official" implementation
of HttpKernelBrowser.

After the Symfony migration we should use
[Rector](https://github.com/rectorphp/rector) with the
RemoveUnusedAliasRector rule to remove the `Client` alias.