Make mailer template path absolute
Before the update to Twig 3.0 the environment configurator made the
paths absolute, now that responsibility is on the FunFunFactory which
did not make the mailer path absolute.

Having path setters on the FunFunFactory is a sub-optimal solution, but
a "proper" solution (e.g. placeholders) still needs to be discussed.
See https://phabricator.wikimedia.org/T267159 for a suggestion.