Turn off source tracking
The "source" value was too long for HTTP referrers coming from some web
mailers, leading to unneccessary length validation errors. Since we
don't need to track the source anyway (See
https://phabricator.wikimedia.org/T134327 ), we can safely set the
parameter to an empty string on every request.

The ReferrerGeneralizer class will then set the default value from the
configuration. I've changed the default in the configuration to an empty
string and allow empty strings in the schema.

Cut down the mapping table for ReferrerGeneralizer to just one entry,
to save memory. The mapping will never occur, since the value is always
empty.