Move HTTP rate limiting to its own class
Donation and membership controllers limit the amount of
donations/memberships to a certain configured amount, to prevent spam.
They use cookies to track the last time someone has donated/applied for
membership.

The code was identical, but had a flaw: The execution halted with a
fatal error when the time stamp did not have a valid format. The fix
would have meant more duplication, so I extracted the functionality into
a separate class.

This change also sets the cookie for donations earlier - when controller
generates the success response (redirect to payment provider or success
page) instead of misusing the confirmation controller for that
functionality. This ensures that people can't do multiple bogus
donations with external payment providers.