Change SubmissionRateLimit storage to Session
Store the submission rate limit in Session instead of cookie, to avoid
having to parse time formats from cookie and to avoid sending extraneous
data to user.

Added session manipulation methods to WebRouteTestCase.
Adding session access to WebRouteTestCase means the Silex Application
object will be persisted, leading to increased memory usage during unit
tests. To alleviate that, we increase the PHP memory limit in the
Makefile.

Previously, AddDonationRouteTest and ApplyForMembershipRouteTest checked
the rate limit cookie for the correct cookie parameters. I've removed
that check for ApplyForMembershipRouteTest and added a generalized check
to AddDonationRouteTest. This test is skipped for now, because it will
fail until we use CookieBuilder everywhere. This change is already on
the horizon with https://github.com/wmde/FundraisingFrontend/pull/1936
When that that PR is merged, we should remove the `markTestAsSkipped`
instruction and try again.

In general AddDonationRouteTest is a bad place to ensure all cookies
have the right parameters, we only (mis)use that test because we're
fairly sure that it is the route that sets the most cookies.

This is for https://phabricator.wikimedia.org/T268766