Remove deprecated source tracking
We no longer track the "Source" (web page) of a donation inside the
donation application (we use Matomo instead). This pull request removes
the parameters and object constructors for source tracking.

It also improves the layer separation by removing the reliance on the
`DonationTrackingInfo` Domain class for passing impression counts from
controller to presenter (without involving the use case). Instead, we
use the new `ImpressionCounts` value object.

This is for https://phabricator.wikimedia.org/T134327