Remove Symfony Filesystem dependency
It was only used to check for the existence of files, but has no methods
for getting file metadata, which we need for better cache ID generation
(will be done in a separate pull request).

In our tests we use vfsStream file system. This simplifies the tests
because we we don't have to configure the file system structure with
`Filesystem` and file contents with mocked `FileReader` implementations.
Instead, we can use the SimpleFileReader, a wrapper around
`file_get_contents`.