Re-Introduce custom event handler
Change the `HandleExceptions` class to be compatible with Symfony and
reactivated it in services.
Change error presenters to accept Throwable instead of exception.

The value of `HandleExceptions::PRIORITY` is still valid and will lead to
exceptions being picked up by the Symfony Exception logger (which has
priority 0), but has a higher priority than the builtin error renderers
(priorities -64 and -128).