Improve configuration modification in tests
Create TestEnvironmentBootstrapper that can modify the configuration
that's passed to FunFunFactory. This reduces the duplication in
WebRouteTestCase (which was almost-but not quite duplicating the factory
function of EnvironmentBootstrapper).

Other tests descending from KernelTest that still use `TestEnvironment`,
can now implement similar functionality without needing
`TestEnvironment`.  `TestEnvironment` was previously responsible for
changing the config, but is on its way out because it's not Symfony
container-aware.