Remove dependencies on TestEnvironment
TestEnvironment was used to bootstrap the FunFunFactory and rebuild the
database schema. Bootstrapping now takes place in KernelTestCase, using
the container to get the Factory instance, configured for the test (with
possible modifications of the configuration).

Rebuilding the database schema now is a trait that can be included in
any test.

Removed WebRouteTestSetupTest because we don't want to test our tests.