Improve logging configuration
The Logger now uses a `fingers_crossed` handler for uat and production
to avoid logging unneccessary information. When the handler triggers, it
will log both to errbit and a local file. The local file will contain
also all "info" level messages (can be useful to detect which route a
request came from). Errbit will log only notice level and up (because
it's our "alerting" and error aggregation system).

Removed SupportHandler and unified ErrbitLoggerFactory: The Monolog
FallbackGroupHandler does the same job as the SupportHandler and the
monolog configuration is suited better to define nesting and log levels.