On setup, use pre-built assets
Add `download-assets` task to Makefile to copy pre-built assets into the
web/skins directory

Remove reference to external asset host - this is a task you have to do
in your custom configuration.

This change means we can have a seamless "out of the box" experience for
the project, without having to start a separate assets server.

Reinstated the BUILD_DIR variable in the Makefile since it's no longer
included from the skins Makefile. This fixes the `stan` task.